Added comparison and set block
I have created the subsystem building_ee and implemented it in the comparison model. It's clear the capacities are not right but now it's significantly easier to make modifications. My interface is different from his in the fact it directly takes the difference between Ph and Pc, that can be easily changed by moving the sum inside the block but it doubles the inputs, I'm not sure it's worth it. The inputs should be in the same order.